sdk: Remove now-unused dependencies (#4568)
* sdk: Remove now-unused dependencies

#### Problem

Almost all of the code in SDK has been moved out to component crates,
but the SDK still declares direct dependencies on crates that it no
longer uses directly.

#### Summary of changes

Clean things up and remove unused dependencies.

* Add back bs58
